+++ rpcserver/jsonresult/voting.go
@@ -1,10 +1,10 @@
 package jsonresult
 
 type GetEncryptionFlagResult struct {
-	DCBFlag uint32 `json:"DCBFlag"`
-	GOVFlag uint32 `json:"GOVFlag"`
+	DCBFlag byte `json:"DCBFlag"`
+	GOVFlag byte `json:"GOVFlag"`
 }
 
 type GetEncryptionLastBlockHeightResult struct {
-	BlockHeight uint32 `json:"blockHeight"`
+	BlockHeight uint64 `json:"blockHeight"`
 }
